frontend: Flush cached comments when doing a comment list
Also move the fetch and pending queues into store module so we
can clear them without creating a cyclic dependency.
